#3 refactor: theme 패키지 경로 수정
