+++ app/src/main/java/org/sopt/and/core/designsystem/theme/Type.kt
@@ -1,4 +1,4 @@
-package org.sopt.and.ui.theme
+package org.sopt.and.core.designsystem.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
